refactor: address Copilot review feedback
- Extract shared parseSchema() to src/parse-schema.ts (was duplicated)
- Fix Mode output that could print "undefined" for optional conversionMode
- Type update params as DocumentTypeUpdateParams instead of Record<string, unknown>
- Use expect.objectContaining() in create test for resilience
